completed callable climate model entrypoint. updated physics model to predict either temp, sea level, or both. made more progress on the ML model using onnxruntime (still incomplete). integrated GameManager wiht the climate model, with functionality to call the entry method to the climate model prediction method.

+++ Assets/Scripts/DataClasses/HexData/WorldState.cs
@@ -6,9 +6,9 @@ namespace TTT.DataClasses.HexData
     [Serializable]
     public struct WorldState
     {
-        public int Pollution;
-        public int SeaLevel;
-        public int Temp;
+        public float Pollution;
+        public float SeaLevel;
+        public float Temp;
         public int Year;
     }
 }
